Fix sonar issue on PostgreSQLTextBitUtils
