+++ db-protocol/postgresql/src/main/java/org/apache/shardingsphere/db/protocol/postgresql/packet/command/query/extended/bind/protocol/util/PostgreSQLTextBoolUtils.java
@@ -15,21 +15,24 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.db.protocol.postgresql.packet.command.query.extended.bind.protocol;
+package org.apache.shardingsphere.db.protocol.postgresql.packet.command.query.extended.bind.protocol.util;
 
-import org.junit.jupiter.api.Test;
-import org.postgresql.util.PGobject;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
 
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-class PostgreSQLTextJsonUtilsTest {
+/**
+ * Text Bool utility class of PostgreSQL.
+ */
+@NoArgsConstructor(access = AccessLevel.PRIVATE)
+public class PostgreSQLTextBoolUtils {
     
-    @Test
-    void assertParse() {
-        String textValue = "['input']";
-        PGobject actual = PostgreSQLTextJsonUtils.parse(textValue);
-        assertThat(actual.getType(), is("json"));
-        assertThat(actual.getValue(), is(textValue));
+    /**
+     * Get Boolean Text value in PostgreSQL text format.
+     * 
+     * @param jdbcBoolValue bool value for jdbc
+     * @return boolean value in PostgreSQL text format
+     */
+    public static String getTextValue(final Object jdbcBoolValue) {
+        return null == jdbcBoolValue ? null : (Boolean) jdbcBoolValue ? "t" : "f";
     }
 }
